changed some formatting and organization to be mre consistent
