+++ lib/main.dart
@@ -1,52 +1,9 @@
 import 'package:flutter/material.dart';
-import 'SummaryAnswers.dart';
-import 'StoryPage.dart';
+import 'story_page.dart';
+import 'home_page.dart';
 
 void main() {
-  runApp(const StoryPage(title: 'Story Page',));
-}
-
-class SettingsButton extends StatelessWidget
-{
-  const SettingsButton({super.key});
-
-  @override
-  Widget build(BuildContext context)
-  {
-    return TextButton(
-      onPressed: (){},
-      child: const Text("Settings"),
-    );
-  }
-}
-
-class ContinueButton extends StatelessWidget
-{
-  const ContinueButton({super.key});
-
-  @override
-  Widget build(BuildContext context)
-  {
-    return TextButton(
-      onPressed: (){},
-      child: const Text("Continue"),
-    );
-  }
-}
-
-class NewGameButton extends StatelessWidget
-{
-  const NewGameButton({super.key});
-
-  @override
-  Widget build(BuildContext context)
-  {
-    return TextButton
-      (
-      onPressed: (){},
-      child: const Text("New Game"),
-    );
-  }
+  runApp(const MyHomePage(title: "Home Page"));
 }
 
 class MyApp extends StatelessWidget {
@@ -59,33 +16,5 @@ class MyApp extends StatelessWidget {
       home: Scaffold(
       )
     );
-  }
-}
-
-class MyHomePage extends StatelessWidget {
-  const MyHomePage({super.key, required this.title});
-
-  final String title;
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-        home: Scaffold(
-          body: Center(
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: const <Widget>[
-                SettingsButton(),
-                Text(
-                  'Title goes here',
-                ),
-                ContinueButton(),
-                NewGameButton()
-              ],
-            ),
-          ),
-        )
-    );
-
   }
 }
